Tema2/Naranja/Class_Prob/pyproject.toml
[project]
name = "class-prob"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "ruamel-yaml (>=0.19.1,<0.20.0)",
    "nltk (>=3.9.3,<4.0.0)"
]

[tool.poetry]
packages = [{include = "class_prob", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
